% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn_sas.R
\name{conn_sas}
\alias{conn_sas}
\title{conn_sas will create a SAS connection}
\usage{
conn_sas(username = Sys.getenv("USER"), password, domain, host, port)
}
\arguments{
\item{username}{(string, optional) Name of the user to login to the SAS
server. Default = Sys.getenv("USER")}

\item{password}{(string) Password for the user.}

\item{domain}{(string) domain of user for SAS connection}

\item{host}{(string) Address to the SAS server.}

\item{port}{(Integer) Port to the SAS Server}
}
\value{
Returns a \code{SASConnection} object
}
\description{
This function allows you to create a SAS workspace using rJava
  and the SAS Java IOM libraries. Once you have the workspace object, you are
  able to execute SAS code, read logs, and retrieve results.
}
\examples{
\dontrun{
sas <- conn_sas()
}
}
\author{
Harrison McGehee  <harrison.mcgehee@unos.org>
}
