% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn_sas.R
\name{sas_disconnect}
\alias{sas_disconnect}
\title{sas_disconnect will disconnect a SAS connection}
\usage{
sas_disconnect(sas_conn)
}
\arguments{
\item{sas_conn}{(\code{SASConnection}) The connection you want to close.}
}
\value{
Returns a \code{boolean} of whether it succeeded or failed.
}
\description{
This function allows you to disconnect a SAS workspace using
  rJava and the SAS Java IOM libraries. Once you have the workspace object,
  you need to disconnect it from the SAS server after you are done.
}
\examples{
\dontrun{
sas_con <- conn_sas()
sas_disconnect(sas_con)
}
}
\author{
Harrison McGehee  <harrison.mcgehee@unos.org>
}
